Rename Translations.js
